solve boj1546

+++ boj/Boj11720_B4.kt
@@ -5,7 +5,7 @@ import java.io.InputStreamReader
 
 fun main() {
     val br = BufferedReader(InputStreamReader(System.`in`))
-    val N = br.readLine()
+    val N = br.readLine().toInt()
 
     val numbers = br.readLine()
         .toCharArray()
